options panel for color chooser dialog
